Fix #142006: Color Jitter produces black points in Material mode
This fixes a bug caused by 96e549c092. Before the change, the
`randomize_color` function passed through the `color.a` value in the
returned rgba color. The alpha is `1` when the color mode is Color
Attribute and `0` when it's Material. After the patch, it was hard
coded to `1`. This fixes it by passing the `color.a` back through
again.

Pull Request: https://projects.blender.org/blender/blender/pulls/142642