Nodes: support interface items lookup by Identifier
Currently, the `NodeTreeInterface` only allows lookup via the item names, and not
their identifiers. This is problematic when there are multiple items with the
same name, they can only be selected via index and not name / identifier.

The following examples are done on a default Geometry Nodes tree that just has a
`Geometry` input and `Geometry` output.

Pull Request: https://projects.blender.org/blender/blender/pulls/140196