Fix: VSE: Copy to selected would interfere with dual handles
After #139075, since dual handle translation always selects two strips
at a time, any later handle movement on one of the strips would affect
both, which is unintuitive and likely not what the user intended.

We can fix this for now by checking for strips that directly border one
another and removing them from the `copy_to` list.

But in the future, if we change selection logic so that strip and handle
selection states are mutually incompatible, we could also fix this by
making sure that dual handle tweaking only selects the handles and not
the strips themselves.

Pull Request: https://projects.blender.org/blender/blender/pulls/142916