Fix #141732: Asset Screenshot does not capture for Cycles
When trying to capture an asset preview screenshot of the 3D viewport while in
Cycles render preview mode, the screenshot would not capture any geometry.

It turns out the function used (`ED_view3d_draw_offscreen_imbuf`) is not useful
for cycles because it doesn't give cycles time to render.

Instead of having cycles rendering, potentially taking a long time, we fall back to
a direct screencapture.
This has the disadvantage that the background will not be transparent in this case,
but this is a limitation we can document.
Also, instead of only excluding Cycles, I chose to include only EEVEE. That way,
things will work even when using 3rd party render engines.

Pull Request: https://projects.blender.org/blender/blender/pulls/142241