Fix #134231: Allow Use of Shortcuts to External Operation Items
Right-clicking on a File Browser item shows a context menu that
(usually) includes a submenu "External". The items on that work
correctly when selected from the menu, but will not work correctly if
you create a keyboard shortcut directly to the item. This PR removes
the use of a filepath property on these, instead having each operate
on the currently-selected File Browser item.

Pull Request: https://projects.blender.org/blender/blender/pulls/134819