Fix #143148: Attribute shader node no longer finds custom property
This code was not meant to be changed to system ID properties, it was
explicitly looking for custom properties before resolving the RNA path.

There are tests for this functionality, but because the versioning copies
the ID properties to the system ID Properties, it worked for existing
properties but not newly created ones.

Pull Request: https://projects.blender.org/blender/blender/pulls/143574